test(team_crud): rebase and refactor packages name
